src/domain/forum/application/use-cases/best-answer-use-case.spec.ts
import { InMemoryAnswersRepository } from 'test/repositories/in-memory-answers-repository';
import { makeAnswer } from 'test/factories/make-answer';
import { UniqueEntityID } from '@/core/entities/unique-entity-id';
import { InMemoryQuestionsRepository } from 'test/repositories/in-memory-questions-repository';
import { BestAnswerQuestionUseCase } from './best-answer-use-case';
import { makeQuestion } from 'test/factories/make-question';

let inMemoryAnswersRepository: InMemoryAnswersRepository;
let inMemoryQuestionsRepository: InMemoryQuestionsRepository;
// system under test
let sut: BestAnswerQuestionUseCase;

describe('Choose question best answer', () => {
  beforeEach(() => {
    inMemoryAnswersRepository = new InMemoryAnswersRepository();
    inMemoryQuestionsRepository = new InMemoryQuestionsRepository();

    sut = new BestAnswerQuestionUseCase(inMemoryAnswersRepository, inMemoryQuestionsRepository);
  });
  it('should be able to choose question best answer', async () => {
    const newQuestion = makeQuestion({});
    const newAnswer = makeAnswer({
      questionId: newQuestion.id,
    });

    await inMemoryQuestionsRepository.create(newQuestion);
    await inMemoryAnswersRepository.create(newAnswer);

    // Prefered way
    await sut.execute({
      answerId: newAnswer.id.toString(),
      authorId: newQuestion.authorId.toString(),
    });
    // Alternative way
    // const { answer } = await sut.execute({ slug: 'example-answer' });
    expect(inMemoryQuestionsRepository.items.get(newQuestion.id)?.bestAnswerId).toEqual(
      newAnswer.id
    );
  });

  it('should not be able to another user question best answer', async () => {
    const newQuestion = makeQuestion({
      authorId: new UniqueEntityID('author-1'),
    });
    const newAnswer = makeAnswer({
      questionId: newQuestion.id,
    });

    await inMemoryQuestionsRepository.create(newQuestion);
    await inMemoryAnswersRepository.create(newAnswer);
    // Prefered way
    // await sut.execute({ authorId: 'author-2', answerId: 'answer-1' });
    // const a = await sut.execute({ answerId: 'answer-1' });
    // Alternative way
    // const { answer } = await sut.execute({ slug: 'example-answer' });
    expect(() => {
      return sut.execute({ authorId: 'author-2', answerId: newAnswer.id.toString() });
    }).rejects.toBeInstanceOf(Error);
  });

  it('should not be able to choose best answer from an invalid answer', async () => {
    const newAnswer = makeAnswer(
      {
        authorId: new UniqueEntityID('author-1'),
      },
      new UniqueEntityID('answer-1')
    );
    await inMemoryAnswersRepository.create(newAnswer);
    // Prefered way
    // await sut.execute({ authorId: 'author-2', answerId: 'answer-1' });
    // const a = await sut.execute({ answerId: 'answer-1' });
    // Alternative way
    // const { answer } = await sut.execute({ slug: 'example-answer' });
    expect(() => {
      return sut.execute({ authorId: 'author-1', answerId: 'answer-2' });
    }).rejects.toBeInstanceOf(Error);
  });
});
